Extract browser portal omnibar suggestion views into their own file
Pure code motion ahead of the #7380 fix: Sources/BrowserWindowPortal.swift
sits at its file-length budget, so the omnibar suggestions overlay types
move verbatim (with internal visibility for test access) into
Sources/BrowserPortalOmnibarSuggestions.swift. No behavior change.
